pyproject.toml
[project]
name = "openenv-bug-triage-env"
version = "1.0.0"
description = "Bug triage simulation environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["environment*", "models*", "utils*", "graders*", "tasks*", "core*"]
